made jit calls more lenient (#224)
I was forcing nopython = True on jitted functions which causes errors whenever a function can't be compiled in that mode for any reason. With that removed it should fail compiling and then fallback to python mode which is better than the bot not functioning at all. So performance gains are maintained on computers than can compile properly and at least diminished functionality is permitted to those that can't.